Private Game stuff (Read Description)
Add message to inform players that the game will start when a minimum
player amount is reached.
Temporary fix for games stopping when minimum players hit.  This is
because private lobbies will work with a simple 4 player minimum / 4
player maximum to lockdown a lobby for 4 players. If one leaves it would
end the game, which we don't want. All public maps have a minimum of 1
player anyways. This can be re-worked if needed.